requirements.txt
pdfminer.six
PyMuPDF
camelot-py[cv]
ghostscript
pandas
python-docx

